# perf(dotnet): fail faster in the dotnet module