@davidkna 2020-03-28
It looks like #886 wasn't fixed after all.