@matchai 2020-04-03
Thank you for the performance improvement, @davidkna! 😄